+++ backend/src/baserow/contrib/database/config.py
@@ -14,12 +14,14 @@ class DatabaseConfig(AppConfig):
         plugin_registry.register(DatabasePlugin())
 
         from .fields.field_types import (
-            TextFieldType, LongTextFieldType, NumberFieldType, BooleanFieldType
+            TextFieldType, LongTextFieldType, NumberFieldType, BooleanFieldType,
+            DateFieldType
         )
         field_type_registry.register(TextFieldType())
         field_type_registry.register(LongTextFieldType())
         field_type_registry.register(NumberFieldType())
         field_type_registry.register(BooleanFieldType())
+        field_type_registry.register(DateFieldType())
 
         from .views.view_types import GridViewType
         view_type_registry.register(GridViewType())